% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\name{arframer_dependency}
\alias{arframer_dependency}
\alias{ar_scene}
\alias{embed_arframer}
\alias{ar_marker_camera}
\alias{a_marker_camera}
\title{Functions}
\usage{
arframer_dependency(cdn = FALSE)

ar_scene(..., cdn = FALSE)

embed_arframer(ar)

ar_marker_camera(...)

a_marker_camera(...)
}
\arguments{
\item{cdn}{Wether to use local of CDN for source files.}

\item{...}{Any \code{aframer} element or option.}

\item{ar}{And \code{arframer} scene.}
}
\description{
Attach \code{aframer} and \code{arframer} dependency, primitives and embed.
}
\examples{
# use the primitive
ar_scene()

# OR
library(aframer)

a_scene(
  a_dependency(), # aframe dependency
  arframer_dependency(), # arframer dependency
  arjs = NA
)

}
